==> 02_explore_gbif.Rmd <==
---
title: "Exploring GBIF"
output: html_document
---
# Background
The Global Biodiversity Information Facility (GBIF) is the largest repository for geo-referenced location data for species across the tree of life, currently providing access to more than 1.3 billion occurrence records.


# Objectives
After this exercise students will have a idea of the data that is available via GBIF and the metadata describing it.

# Exercise
Use the GBIF webpage (www.gbif.org) to explore what kind of data is available and what kind of meta-data comes with it.

- Which is the taxon with most records in GBIF?

- When was the first record collected in Germany?

- What are the major contributing databases in Germany?

- What are the major basis of record in GBIF?

- How many records of plants were added from above 6000 meters elevation?

- How many records are there for your taxon of interest and which is it?
